requirements.txt
flask>=2.3.0
openai-whisper>=20231117
yt-dlp>=2023.11.16
openai>=1.3.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
gunicorn>=21.2.0
